pyproject.toml
[project]
name = "assistant-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
dev = [
    "google-genai>=1.12.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.3",
    "pydantic-ai>=0.1.8",
    "pydantic-settings>=2.9.1",
]
gcp = [
    "gcsfs>=2025.3.0",
    "google-auth>=2.38.0",
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-iam>=2.19.0",
    "google-cloud-run>=0.10.17",
    "google-cloud-secret-manager>=2.23.2",
    "google-cloud-storage>=3.1.0",
]
notebooks = [
    "dotenv>=0.9.9",
    "ipykernel>=6.29.5",
]
